requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
fastapi==0.85.2
fastapi-oauth==0.3.4
fastapi-sso==0.6.4
fire==0.5.0
frozenlist==1.3.3
github==1.2.7
google==3.0.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
Jinja2==3.1.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.8
packaging==23.1
passlib==1.7.4
pluggy==1.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
pyparsing==3.0.9
pytest==7.3.2
pytest-mock==3.10.0
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
ruff==0.0.272
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.48
starlette==0.20.4
stripe==5.4.0
termcolor==2.3.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.22.0
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.15.0